ant_plus_sensor/requirements.txt
ant==0.1.1
paho-mqtt==1.6.1